completed all exercises of part 4

+++ part4/bloglist/controllers/blogRouter.js
@@ -3,6 +3,7 @@ const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
 const User = require("../models/user")
 const jwt = require("jsonwebtoken")
+const { userExtractor } = require("../utils/middleware")
 
 blogsRouter.get("/", async (request, response) => {
     const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
@@ -16,7 +17,7 @@ blogsRouter.get("/:id", async (request, response) => {
         : response.status(404).end()
 })
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
     const decodedToken = jwt.verify(request.token, config.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({ error: "token missing or invalid" })
@@ -35,9 +36,13 @@ blogsRouter.post("/", async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
     const user = request.user
     const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).end()
+    }
 
     if (blog.user.toString() === user.id.toString()) {
         await Blog.findByIdAndDelete(request.params.id)
