+++ part4/bloglist/tests/blog_api.test.js
@@ -5,6 +5,9 @@ const app = require("../app")
 const api = supertest(app)
 const Blog = require("../models/blog")
 const User = require("../models/user")
+const jwt = require("jsonwebtoken")
+const bcrypt = require("bcrypt")
+const config = require("../utils/config")
 
 beforeEach(async () => {
   await Blog.deleteMany({})
@@ -24,6 +27,16 @@ describe("Testing GET requests", () => {
 })
 
 describe("Testing POST requests", () => {
+  let token = null
+  beforeAll(async () => {
+    await User.deleteMany({})
+
+    const passwordHash = await bcrypt.hash("98765", 10)
+    const user = await new User({ username: "test", passwordHash }).save()
+    const userToken = { username: "test", id: user.id }
+    return (token = jwt.sign(userToken, config.SECRET))
+  })
+
 
   test("a valid blog can be added", async () => {
     const newBlog = {
@@ -33,7 +46,7 @@ describe("Testing POST requests", () => {
       likes: 2
     }
 
-    await api.post("/api/blogs").send(newBlog).expect(201).expect("Content-Type", /application\/json/)
+    await api.post("/api/blogs").set("Authorization", `bearer ${token}`).send(newBlog).expect(201).expect("Content-Type", /application\/json/)
 
     const blogsAtEnd = await helper.blogsInDb()
     expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length + 1)
@@ -49,7 +62,7 @@ describe("Testing POST requests", () => {
       url: "www.example.com"
     }
 
-    await api.post("/api/blogs").send(newBlog).expect(201).expect("Content-Type", /application\/json/)
+    await api.post("/api/blogs").send(newBlog).set("Authorization", `bearer ${token}`).expect(201).expect("Content-Type", /application\/json/)
 
     const blogsAtEnd = await helper.blogsInDb()
     expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length + 1)
@@ -63,20 +76,68 @@ describe("Testing POST requests", () => {
       author: "Dom"
     }
 
-    await api.post("/api/blogs").send(newBlog).expect(400)
+    await api.post("/api/blogs").set("Authorization", `bearer ${token}`).send(newBlog).expect(400)
     const blogsAtEnd = await helper.blogsInDb()
     expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length)
   })
 })
 
 describe("Testing DELETE method", () => {
-  test("test succeeds with status 204", async () => {
+  beforeEach(async () => {
+    await User.deleteMany({})
+
+    const passwordHash = await bcrypt.hash("98765", 10)
+    const user = await new User({ username: "test", name: "Test", passwordHash }).save()
+    await Blog.deleteMany({})
+
+    for (let blog of helper.initialBlogs) {
+      let blogObject = new Blog(blog)
+      blogObject.user = user._id
+      await blogObject.save()
+      user.blogs = user.blogs.concat(blogObject._id.toString())
+      await user.save()
+    }
+  })
+
+  const loginToken = async () => {
+    const userLoginDetails = {
+      username: "test",
+      password: "98765"
+    }
+
+    const login = await api.post("/api/login").send(userLoginDetails).expect(200)
+    const token = login.body.token
+    expect(token).not.toBe(null)
+    return token
+  }
+
+  test("test succeeds when post is deleted by user who created it", async () => {
+    const token = await loginToken()
     const blogsAtStart = await helper.blogsInDb()
     const blogToDelete = blogsAtStart[0]
-    await api.delete(`/api/blogs/${blogToDelete.id}`).expect(204)
+    const userBlogsAtStart = (await helper.usersInDb()).find(user => user.username === "test").blogs.map(blog => blog._id.toString())
+    expect(userBlogsAtStart).toContain(blogToDelete.id)
+
+    await api.delete(`/api/blogs/${blogToDelete.id}`).set("Authorization", `bearer ${token}`).expect(204)
+
     const blogsAtEnd = await helper.blogsInDb()
     expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length - 1)
-    expect(blogsAtEnd).not.toContainEqual(blogToDelete)
+
+    const titles = blogsAtEnd.map(b => b.title)
+    expect(titles).not.toContain(blogToDelete.title)
+  })
+
+  test("test succeeds with 401 unauthorized action user tries to delete another user's blog", async () => {
+    const token = null
+    const blogsAtStart = await helper.blogsInDb()
+    const blogToDelete = blogsAtStart[0]
+    const userBlogsAtStart = (await helper.usersInDb()).find(user => user.username === "test").blogs.map(blog => blog._id.toString())
+    expect(userBlogsAtStart).toContain(blogToDelete.id)
+
+    await api.delete(`/api/blogs/${blogToDelete.id}`).set("Authorization", `bearer ${token}`).expect(401)
+
+    const blogsAtEnd = await helper.blogsInDb()
+    expect(blogsAtEnd).toHaveLength(blogsAtStart.length)
   })
 })
